2 changes here (contentSlider.jquery.js):
a.) in the case there's NOT more than 1 slide, don't do most of the setup functions, like AttachControlBar(), DoMiscSetup(), and StartShow(). We still run ApplyCSS() though.
b.) fixed glitch with mouseover behavior. Needed to run StopShow(), and then in the timeout function that polls to see if the mouseover condition is false yet, and if so, then run StartShow().

And then as for images/indicator-sprite-20x40.build.png, this is an updated graphic, looks a little nicer than the original ones... much smaller.